an api folder was created and the logic was transferred to it
