+++ .env_example
@@ -1,0 +1,1 @@
+VITE_API_KEY=my_vite_api_key
