+++ src/mapServices.ts
@@ -3,26 +3,26 @@ import 'leaflet/dist/leaflet.css';
 
 import { Airport, AirportAirQuality } from './store/types';
 
-const tileLayer = L.tileLayer('https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png', 
+const tileLayer = L.tileLayer('https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
   {
     attribution: 'Map data &copy; <a href="http://openstreetmap.org">OpenStreetMap</a> contributors'
-});
+  });
 
 export default class MapService {
   private mapContainer: HTMLElement;
   private map: L.Map;
 
   constructor(
-    mapContainer: HTMLElement, 
-    center: [number, number] = [48.8566, 2.3522], 
+    mapContainer: HTMLElement,
+    center: [number, number] = [48.8566, 2.3522],
     zoom: number = 4) {
-      this.mapContainer = mapContainer;
-      this.map = L.map(this.mapContainer).setView(center, zoom);
+    this.mapContainer = mapContainer;
+    this.map = L.map(this.mapContainer).setView(center, zoom);
     tileLayer.addTo(this.map);
   }
 
   addMarkers(airports: Airport[]) {
-    if (!this.map) return; 
+    if (!this.map) return;
     airports.forEach((airport) => {
       const { latitude, longitude, name, icao } = airport;
       const marker = L.marker([latitude, longitude]).addTo(this.map);
@@ -34,18 +34,18 @@ export default class MapService {
     });
   }
 
-  addAirQualityMarkers(airports: Airport[], AirportAirQuality:Record<string, AirportAirQuality>){
-    if (!this.map)return;
+  addAirQualityMarkers(airports: Airport[], AirportAirQuality: Record<string, AirportAirQuality>) {
+    if (!this.map) return;
     airports.forEach((airport) => {
       const { latitude, longitude, icao } = airport;
       const marker = L.marker([latitude, longitude]).addTo(this.map);
       const airQuality = AirportAirQuality[icao];
-      const airQualityInfo = airQuality 
+      const airQualityInfo = airQuality
         ? `Показатели воздуха: 
         SO2: ${airQuality.SO2.concentration}, 
         NO2: ${airQuality.NO2.concentration}, 
         CO: ${airQuality.CO.concentration}, 
-        O3: ${airQuality.O3.concentration}` 
+        O3: ${airQuality.O3.concentration}`
         : 'Air Quality Data Not Available';
       marker.bindTooltip(airQualityInfo, { permanent: false });
     })
